requirements.txt
PdfReader
pillow
PyPDF2
pdf2image
reportlab
urllib3==1.25.11
pyshorteners
python-decouple